I used a tool called smtp4dev to test the mail feature. Ideally instead of constructing a client and passing it "localhost", you would leave this blank and add credentials in the following lines. At least that was the way i understood it.
============================================================

the regular expression i used to validate the email address was taken from here:
http://www.regular-expressions.info/email.html

I took the second expression presented in the section titled:
The Official Standard: RFC 2822

=======================================================

I still need to add more formatting to the css for this page.
I've been working on this most of the day and I had another class and needed to stop,
so I went ahead and committed it so that at least something would be on the server.